requirements.txt
aioboto3==11.2.0
aiobotocore==2.5.0
aiohttp==3.8.4
aioitertools==0.11.0
aiosignal==1.3.1
anyio==3.7.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
astroid==2.13.5
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
black==22.12.0
bleach==6.0.0
blinker==1.6.2
boto3==1.26.76
botocore==1.29.76
certifi==2023.5.7
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==3.1.0
click==8.1.3
comm==0.1.3
cryptography==41.0.1
dataclasses-json==0.5.8
debugpy==1.6.7
decorator==5.1.1
deeplake==3.6.6
defusedxml==0.7.1
dill==0.3.6
distlib==0.3.6
entrypoints==0.4
exceptiongroup==1.1.1
executing==1.2.0
fastjsonschema==2.17.1
filelock==3.12.2
Flask==2.3.2
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.6.0
greenlet==2.0.2
humbug==0.3.1
identify==2.5.24
idna==3.4
iniconfig==2.0.0
ipykernel==6.23.3
ipython==8.10.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
isort==5.12.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonpointer==2.4
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
langchain==0.0.216
langchainplus-sdk==0.0.17
lazy-object-proxy==1.9.0
-e git+https://github.com/jerryjliu/llama_index.git@d9bdd8c55f29837d41b037df41799a5765cdd20d#egg=llama_index
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib-inline==0.1.6
mccabe==0.7.0
mistune==3.0.1
multidict==6.0.4
multiprocess==0.70.14
mypy==0.991
mypy-extensions==1.0.0
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.6.0
nbformat==5.9.0
nest-asyncio==1.5.6
nltk==3.8.1
nodeenv==1.8.0
notebook==6.5.4
notebook_shim==0.2.3
numcodecs==0.11.0
numexpr==2.8.4
numpy==1.25.0
openai==0.27.8
openapi-schema-pydantic==1.2.4
overrides==7.3.1
packaging==23.1
pandas==2.0.2
pandocfilters==1.5.0
parso==0.8.3
pathos==0.3.0
pathspec==0.11.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.8.0
pluggy==1.2.0
pox==0.3.2
ppft==1.7.6.6
pre-commit==3.2.0
prometheus-client==0.17.0
prompt-toolkit==3.0.38
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydantic==1.10.9
Pygments==2.15.1
PyJWT==2.7.0
pylint==2.15.10
pyrsistent==0.19.3
pytest==7.2.1
pytest-asyncio==0.21.0
pytest-dotenv==0.5.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
pytz==2023.3
PyYAML==6.0
pyzmq==25.1.0
qtconsole==5.4.3
QtPy==2.3.1
rake-nltk==1.0.6
regex==2023.6.3
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
ruff==0.0.259
s3transfer==0.6.1
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.17
stack-data==0.6.2
tenacity==8.2.2
terminado==0.17.1
tiktoken==0.4.0
tinycss2==1.2.1
tk==0.1.0
tomli==2.0.1
tomlkit==0.11.8
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
types-docutils==0.20.0.1
types-pyOpenSSL==23.2.0.1
types-redis==4.5.5.0
types-requests==2.28.11.8
types-setuptools==67.1.0.0
types-urllib3==1.26.25.13
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2023.3
uri-template==1.3.0
urllib3==1.26.16
virtualenv==20.23.1
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
Werkzeug==2.3.6
widgetsnbextension==4.0.7
wrapt==1.15.0
yarl==1.9.2
